refactor(linebreak): share the stub metrics between unit suites
compile.test.ts carried the only FontMetrics stub in the unit tests: a
segmenter that splits on spaces and U+00AD, ten pixels a character, and
an ExtractedBlock builder with one wrapper. Every test that needs items
compiled from real text needs that stub, and copying it into a second
file would let the two drift.

It moves to unit/support/measure.ts unchanged, mirroring e2e/support.
compile() also takes a LayoutPolicy now, so a caller can tell
hyphenPenalty and exHyphenPenalty apart by giving them different values.
